% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\docType{data}
\name{fish}
\alias{fish}
\title{Time series of fish}
\format{
Time Series with 4552 rows including 883 NAs.
}
\usage{
fish
}
\description{
Time series fishes
}
\details{
The time series is derived from ...
}
\keyword{datasets}
