% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imputations.R
\name{get_imputations}
\alias{get_imputations}
\title{get_imputations function}
\usage{
get_imputations(x_ts, methods = "all", seed = 1234)
}
\arguments{
\item{x_ts}{A ts object with missing data to be imputed}

\item{seed}{Numeric. Any number}

\item{method}{A string or string vector indicating the method or methods}
}
\value{
A list of imputed time series with the selected methods
}
\description{
Function to get imputations from methods compared by kssa
}
\examples{
# Get imputed values for co2_na_ts with the methods of
# your choice. Ideally those previously compared by kssa

my.imputations = get_imputations(co2_na_ts, methods = "all")

# my.imputations contains the imputed time series with all methods.
# Access it and choose the one from the best method for your purposes

my.imputations$seadec
plot.ts(my.imputations$seadec)


}
