% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kssa.R
\name{kssa}
\alias{kssa}
\title{kssa Algorithm}
\usage{
kssa(
  x_ts,
  start_method,
  methods,
  segments,
  iterations,
  percentmd = sample(x = 1:50, size = 1)/100,
  seed = 1234
)
}
\arguments{
\item{x_ts}{Time series object (ts) containing missing data (NA)}

\item{segments}{Into how many segments the time series will be divided}

\item{iterations}{How many iterations to run}

\item{percentmd}{Percent of missing data. Must match with the true percentage of missing data in x_ts}

\item{seed}{Random seed to choose (can be any number)}

\item{start_methods}{The method or methods to start the algorithm.
It can be a string vector containing the following 11 iimputation methods

\itemize{
\item{c("auto.arima", "StructTS", "linear_i", "spline_i", "stine_i",
"simple"_i, "simple_ma", "linear_ma", "exponential_ma", "seadec",
"locf", "decomp")} (default choice) - or a subset of these
\item{set to "all" to start with all methods automatically} - Default choice
}}

\item{actual_methods}{The imputation methods to compare. It can be a string vector containing the following
11 imputation methods:

\itemize{
\item{c("auto.arima", "StructTS", "linear_i", "spline_i", "stine_i",
"simple"_i, "simple_ma", "linear_ma", "exponential_ma", "seadec",
"locf", "decomp")} (default choice) - or a subset of these
\item{or "all" to compare among all methods automatically} - Default choice
}

For further details on these imputation methods please check packages \code{\link{imputeTS}} and \code{\link{forecast}}}
}
\value{
A list of results to be plotted with function \code{\link{kssa_plot}} for easy interpretation
}
\description{
Run the Known Sub-Sequence Algorithm to compare imputation methods for time series data
}
\examples{
# create a numeric vector with 20\% missing data
x = c(1, 5, 6, 8, 4, NA, 5, 4, NA, NA)

# convert x to a time series object
x_ts = ts(x)

# apply the kssa algorithm with 2 segments,
# 10 iterations and 20\% of missing data.
# Remember that percentmd must match with
# the real percentaje of missing data in the
# input time series

results_kssa = kssa(x_ts,
               start_method = "all",
               methods = "all",
               segments = 2,
               iterations = 10,
               percentmd = 0.2)

# print results
results_kssa

# plot complete results with Root Mean Squared Error for easy
# interpretation
kssa_plot(results_kssa, type = "complete", metric = "rmse")

# Conclusion: Since the kssa_plot is ordered from lower to
# higher (left to right) average error, the method
# exponential_ma (exponential moving average) is
# the best to impute missing data in x_tx.

}
\references{
Benavides, I. F., Santacruz, M., Romero-Leiton, J. P., Barreto, C., & Selvaraj, J. J. (2022). Assessing methods for multiple imputation of systematic missing data in marine fisheries time series with a new validation algorithm. Aquaculture and Fisheries.
\href{https://www.sciencedirect.com/science/article/pii/S2468550X21001696}{Full text publication}.
}
