% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kssa_plot.R
\name{kssa_plot}
\alias{kssa_plot}
\title{kssa_plot function}
\usage{
kssa_plot(results, type, metric)
}
\arguments{
\item{results}{A list object with kssa.table objects}

\item{type}{A character with type of plot complete or summary}

\item{metric}{A character with metric choosed to be plot}
}
\value{
The plot of \code{results}
}
\description{
Function to plot kssa results
}
