% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kssa.R
\name{kssa}
\alias{kssa}
\title{kssa Algorithm}
\usage{
kssa(
  x_ts,
  start_method,
  methods,
  segments,
  iterations,
  percentmd = sample(x = 1:50, size = 1)/100,
  seed = 1234
)
}
\arguments{
\item{x_ts}{time series object (ts) containing missing data (NA)}

\item{start_method}{The starting method for the algorithm}

\item{methods}{actual.methods    A string vector containing the following
methods: c(“auto.arima”, “StructTS”, “linear”, “spline”, “stine”,
“simple”, “malinear”, “exponential”, “kalman”, “nalocf”, “decomp”) .
“all”    compares among all imputation methods}

\item{segments}{Into how many segments the dataset shall be divided}

\item{iterations}{How many iterations to run}

\item{percentmd}{Percent of missing data}

\item{seed}{Seed to choose}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Function to add two numbers
}
