% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kssa2.R
\name{kssa2}
\alias{kssa2}
\title{kssa2 Algorithm}
\usage{
kssa2(x_ts, start_method, methods, segments, iterations, seed = 1234)
}
\arguments{
\item{x_ts}{The input time series}

\item{start_method}{The starting method for the algorithm}

\item{methods}{The methods that shall be compared}

\item{segments}{Into how many segments the dataset shall be divided}

\item{iterations}{How many iterations to run}

\item{seed}{Seed to choose}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Function to add two numbers
}
