% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kssa_plot.R
\name{kssa_plot}
\alias{kssa_plot}
\title{kssa_plot function}
\usage{
kssa_plot(results, type, metric)
}
\arguments{
\item{results}{A list object with results produced with function \code{\link{kssa}}}

\item{type}{A character value with the type of plot to show.
It can be "summary" or "complete".}

\item{metric}{A character with the performance metric to be plotted
It can be  "rmse", "mase," "cor", or "smape"

For further details on these metrics please check package \code{\link{Metrics}}}
}
\value{
A plot of kssa results ordered from lower to higher (left to right)
averaged error
}
\description{
Function to plot the results of kssa for easy interpretation
}
\examples{
# create a numeric vector with 20\% missing data
x = c(1, 5, 6, 8, 4, NA, 5, 4, NA, NA)

# convert x to a time series object
x_ts = ts(x)

# apply the kssa algorithm with 2 segments,
# 10 iterations and 20\% of missing data.
# Remember that percentmd must match with
# the real percentaje of missing data in the
# input time series

results_kssa = kssa(x_ts,
               start_method = "all",
               methods = "all",
               segments = 2,
               iterations = 10,
               percentmd = 0.2)

# print results
results_kssa

# plot complete results with Root Mean Squared Error for easy
# interpretation
kssa_plot(results_kssa, type = "complete", metric = "rmse")

# Conclusion: Since the kssa_plot is ordered from lower to
# higher (left to right) average error, the method
# exponential_ma (exponential moving average) is
# the best to impute missing data in x_tx.

}
